Add an option for ignoring undefined field
When I remove a field from a model, I have to remove the corresponding
attribute  from all saved items in DynamoDB. That is for avoiding
`NoMethodError` on loading an item from DB. That may require to stop the
web service during running a script of deleting that attributes. This
commit is a proposal of adding an option to ignore undefined (but
attributes are exist in saved items) fields on loading an item.